% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Returns an answer to a query hash}
\usage{
query(query_hash = first_version_query_hash(),
  url_endpoint = query_internet_archive,
  hash2path = hash_to_2level_path())
}
\value{
an answer in the form a content hash
}
\description{
Internally, an answer is retrieved by appending the relative path
of the query hash provided by hash2path to the url endpoint:
}
\details{
[url endpoint]/[query hash path]

where query hash path = hash2path(query_hash)
}
\examples{
\donttest{
 first_version <- query("hash://sha256/2a5de79372318317a382ea9a2cef069780b852b01210ef59e06b640a3539cb5a")
 # "hash://sha256/c253a5311a20c2fc082bf9bac87a1ec5eb6e4e51ff936e7be20c29c8e77dee55"
}
 @seealso https://github.com/bio-guoda/preston/blob/master/architecture.md#simplified-hexastore#'
 @seealso https://preston.guoda.bio

}
