% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_internet_archive}
\alias{query_internet_archive}
\title{Returns answer to query hash as provided by Preston remote at
https://archive.org/download/biodiversity-dataset-archives/data.zip/data/}
\usage{
query_internet_archive(query_hash = first_version_query_hash())
}
\value{
an answer in the form a content hash
}
\description{
Returns answer to query hash as provided by Preston remote at
https://archive.org/download/biodiversity-dataset-archives/data.zip/data/
}
\examples{
\donttest{
 first_version <- query_deep_linker("hash://sha256/2a5de79372318317a382ea9a2cef069780b852b01210ef59e06b640a3539cb5a")
 # "hash://sha256/c253a5311a20c2fc082bf9bac87a1ec5eb6e4e51ff936e7be20c29c8e77dee55"
}
}
\seealso{
https://archive.org/download/biodiversity-dataset-archives/data.zip/data/
}
