% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Opens a stream (i.e., connection) to retrieve content associated with content_hash}
\usage{
retrieve(content_hash,
  hash2url = hash2url_ia_biodiversity_dataset_archive)
}
\description{
@param content_hash content hash of requested content in hashuri format.
@return stream (i.e., connection) of content
}
\details{
@export
}
