% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_provenance.R
\name{stream_provenance}
\alias{stream_provenance}
\title{Streams processed provenance log into writable connection .}
\usage{
stream_provenance(con, process_func = function(x, version_id,
  line_number_offset) {     x }, version_iter = version_history_iter(),
  n = -1L, line_batch_size = 1)
}
\arguments{
\item{con}{the connection (e.g., file) to write into after being processed by process_func}

\item{process_func}{a function that processes provenance lines}

\item{version_iter}{an provenance version iterator @seealso version_history_iter()}

\item{n}{maximum number of (processed) lines to write, -1 means unlimited}

\item{line_batch_size}{number of lines to be processed at once}
}
\value{
number of written (processed) lines
}
\description{
Streams processed provenance log into writable connection .
}
\examples{
\donttest{
 # open an temporary read/write connection
 con <- fifo("", open = "w+b")
 stream_provenance(con, n = 1)
 first_line_first_provenance_log <- readLines(con)
 close(con)

 # open file
 con <- file("somefile.txt", open = "w+b")
 versions_only = function(lines) {
   lines[grepl("hasVersion", lines)]
 }

 # save only provenance lines with "hasVersion" in it
 stream_provenance(con, process_func = versions_only, n = 1)
 first_version_in_provenance_log <- readLines(con)

 close(con)
}

}
