% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_history.R
\name{version_history}
\alias{version_history}
\title{List preston version history starting with oldest version.}
\usage{
version_history(query = query_internet_archive)
}
\arguments{
\item{query}{a function that takes a query hash and returns a content hash}
}
\value{
list of versions, starting with the oldest
}
\description{
Please use version_history_iter for incrementally retrieving version.
}
\examples{
\donttest{
 version_list <- version_history(query_internet_archive)
 head(version_list)
 # "hash://sha256/c253a5311a20c2fc082bf9bac87a1ec5eb6e4e51ff936e7be20c29c8e77dee55"
}
Learn more about preston at:

  https://preston.guoda.bio

}
